:stada: Update to v2.0.0 Now we are much more than a selector library, we can easyli manage frames and videos
