+++ cv2_tools/__init__.py
@@ -1,13 +1,12 @@
-from opencv_draw_tools.SelectZone import *
-import opencv_draw_tools.tags_constraint
+from cv2_tools.utils import webcam_test, get_complete_help
 
-name = 'opencv_draw_tools'
+name = 'cv2_tools'
 help = '''
 MIT License
 Copyright (c) 2019 Fernando Perez
 For more information visit: https://github.com/fernaper/opencv-draw-tools
 Also you can write complete_help to view full information'''
-__version__ =  '1.2.0'
+__version__ =  '2.0.0'
 
 complete_help = '''
 {} - v{}
