+++ examples/cv2_tools
@@ -1,0 +1,1 @@
+../opencv_draw_tools
