+++ examples/detect_faces.py
@@ -3,19 +3,15 @@
 import face_recognition
 import cv2
 
-try:
-    # If we use this import we are testing the github version
-    import opencv_draw_tools.SelectZone as cv2_tools
-    print('Testing mode')
-except:
-    import opencv_draw_tools as cv2_tools
+from cv2_tools.Selection import  SelectorCV2
+
 
 def face_detector(frame, scale=0.25):
     small_frame = cv2.resize(frame, (0, 0), fx=scale, fy=scale)
     rgb_small_frame = small_frame[:, :, ::-1]
     face_locations = face_recognition.face_locations(rgb_small_frame)
 
-    selector = cv2_tools.SelectorCV2(color=(200,90,0), filled=True)
+    selector = SelectorCV2(color=(200,90,0), filled=True)
     for i, face_location in enumerate(face_locations):
         y1, x2, y2, x1 = [position/scale for position in face_location]
         selector.add_zone((x1,y1,x2,y2),'Face {}'.format(i))
