--- opencv_draw_tools/LICENSE
@@ -1,1 +1,0 @@
-../LICENSE
