--- opencv_draw_tools/README.md
@@ -1,1 +1,0 @@
-../README.md
